requirements.txt
streamlit
google-generativeai
gspread
google-auth          # Library baru pengganti oauth2client
Pillow